+++ sdk/resourcemanager/connectedcache/armconnectedcache/constants.go
@@ -4,11 +4,6 @@
 
 package armconnectedcache
 
-const (
-	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/connectedcache/armconnectedcache"
-	moduleVersion = "v0.1.0"
-)
-
 // ActionType - Extensible enum. Indicates the action type. "Internal" refers to actions that are for internal only APIs.
 type ActionType string