+++ sdk/resourcemanager/containerservicefleet/armcontainerservicefleet/updateruns_client_example_test.go
@@ -6,10 +6,11 @@ package armcontainerservicefleet_test
 
 import (
 	"context"
+	"log"
+
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservicefleet/armcontainerservicefleet/v3"
-	"log"
 )
 
 // Generated from example definition: 2025-04-01-preview/UpdateRuns_CreateOrUpdate.json