+++ sdk/resourcemanager/databasewatcher/armdatabasewatcher/watchers_client.go
@@ -192,7 +192,6 @@ func (client *WatchersClient) deleteCreateRequest(ctx context.Context, resourceG
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2025-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }