+++ sdk/resourcemanager/purestorageblock/armpurestorageblock/constants.go
@@ -4,11 +4,6 @@
 
 package armpurestorageblock
 
-const (
-	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/purestorageblock/armpurestorageblock"
-	moduleVersion = "v1.0.0"
-)
-
 // ActionType - Extensible enum. Indicates the action type. "Internal" refers to actions that are for internal only APIs.
 type ActionType string