+++ sdk/resourcemanager/standbypool/armstandbypool/standbycontainergrouppools_client.go
@@ -193,7 +193,6 @@ func (client *StandbyContainerGroupPoolsClient) deleteCreateRequest(ctx context.
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2025-03-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }